requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.11.0
opencv-python>=4.5.0
jupyter>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
tqdm>=4.62.0
requests>=2.26.0
pillow>=8.3.0
gradio
gtts>=2.2.0